DBFilehandler added

+++ project_2/src/utilities/Agency.java
@@ -1,7 +1,7 @@
 package utilities;
 
 public class Agency extends User {
-	public Agency(String username, String password, String division, int certNbr) {
+	public Agency(String username, String password, String division, long certNbr) {
 		super(username, password, division, certNbr);
 		this.permLevel = PermissionLevel.Agency;
 	}
