+++ project_2/src/utilities/Nurse.java
@@ -1,7 +1,7 @@
 package utilities;
 
 public class Nurse extends User {
-	public Nurse(String username, String password, String division, int certNbr) {
+	public Nurse(String username, String password, String division, long certNbr) {
 		super(username, password, division, certNbr);
 		this.permLevel = PermissionLevel.Nurse;
 	}
